진단데이터 저장 및 유저 사용목적 관련 API 구현

+++ src/main/java/Phonesonal/PhoneBE/web/dto/InfoResponse.java
@@ -4,14 +4,16 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @AllArgsConstructor
 public class InfoResponse {
     private String nickName;
     private String email;
-    private int height;
-    private int weight;
+    private BigDecimal height;
+    private BigDecimal weight;
     private int deadline;
     private long weeksTogether;
 }
